% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{brkpt}
\alias{brkpt}
\title{Fit breakpoint model to individual colony
Fits models using a range of taus and picks the best one using maximum liklihood}
\usage{
brkpt(data, taus, t, formula)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{taus}{a vector of taus to test}

\item{t}{the unquoted variable representing time in ___units?}

\item{formula}{a formula passed to `lm`}
}
\value{
dataframe with two columns
}
\description{
Fit breakpoint model to individual colony
Fits models using a range of taus and picks the best one using maximum liklihood
}
