% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{brkpt2}
\alias{brkpt2}
\title{Same as brkpt but uses glm2
Same as brkpt but uses glm2}
\usage{
brkpt2(data, taus = NULL, t, formula, family = gaussian(link = "log"),
  ...)
}
\description{
Same as brkpt but uses glm2
Same as brkpt but uses glm2
}
