% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bumbl-package.R
\docType{package}
\name{bumbl-package}
\alias{bumbl-package}
\title{bumbl: Tools for Modeling Bumblebee Colony Growth and Decline}
\description{
Bumblebee colonies grow during worker production, then decline after switching to production of reproductive individuals (drones and gynes). This package provides tools for modeling and visualizing this pattern by identifying a switchpoint with a growth rate before and a decline rate after the switchpoint. The mathematical models fit by bumbl are described in Crone and Williams (2016) \doi{10.1111/ele.12581}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Aariq/bumbl}
  \item Report bugs at \url{https://github.com/Aariq/bumbl/issues}
}

}
\author{
\strong{Maintainer}: Eric R. Scott \email{scottericr@gmail.com} (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})

}
\keyword{internal}
