% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{bumbl_plot}
\alias{bumbl_plot}
\title{Plot observed and fitted results from bumbl()}
\usage{
bumbl_plot(bumbldf)
}
\arguments{
\item{bumbldf}{A dataframe produced by \code{bumbl()} with \code{augment = TRUE}}
}
\value{
currently doesn't return an object, just prints plots
}
\description{
Plot observed and fitted results from bumbl()
}
\examples{
colonyID_subset <- sample(bombus$colony, 10)
colony_subset <- bombus[bombus$colony \%in\% colonyID_subset, ]
results <- bumbl(colony_subset, colonyID = colony, t = week,
                 formula = log(mass) ~ week, augment = TRUE)
bumbl_plot(results)
}
