% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{forage_func}
\alias{forage_func}
\title{Probability of foraging as a function of bumblbee worker size}
\usage{
forage_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of foraging probabilities
}
\description{
Probability of foraging as a function of bumblbee worker size
}
\examples{
forage_func(4)
}
\references{
TODO: cite paper this comes from
}
