% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{mass_func}
\alias{mass_func}
\title{Bumblebee worker mass as a function of intertegular span (ITS)}
\usage{
mass_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of worker mass in grams
}
\description{
Bumblebee worker mass as a function of intertegular span (ITS)
}
\examples{
mass_func(4.1)
}
\references{
TODO: cite paper this comes from
}
