% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{poln_mass_func}
\alias{poln_mass_func}
\title{Pollen mass as a function of bumblbee worker size}
\usage{
poln_mass_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of pollen mass in grams
}
\description{
Pollen mass as a function of bumblbee worker size
}
\examples{
poln_mass_func(4)
}
\references{
TODO: cite paper this comes from
}
