% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{poln_ret_func}
\alias{poln_ret_func}
\title{Probability of pollen return as a funciton of bumblebee worker size}
\usage{
poln_ret_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of pollen return probabilities
}
\description{
Probability of pollen return as a funciton of bumblebee worker size
}
\examples{
poln_ret_func(4)
}
\references{
TODO: cite paper this comes from
}
