% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{surv_func}
\alias{surv_func}
\title{Bumblebee worker survival as a function of size}
\usage{
surv_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of survival probabilities
}
\description{
Bumblebee worker survival as a function of size
}
\examples{
surv_func(4)
}
\references{
TODO: cite paper this comes from
}
