% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Simulated colony growth data for internal tests}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140 rows and 3 columns.
}
\usage{
test_df
}
\description{
Colonies 1-5 are generated with \code{sim_colony()} using the following parameters:
\describe{
\item{tau} \code{c(11, 11.5, 9, 12, 10.6)}
\item{n0} \code{c(30, 24, 49, 37, 29)}
\item{lambda} \code{c(1.42, 1.35, 1.4, 1.34, 1.37)}
\item{delta} \code{c(0.3, 0.31, 0.29, 0.48, 0.32)}
}
Colonies 6 and 7 are generated manually.  Colony 6 contains negative numbers
and should result in the warning "Warning, cannot find valid starting values:
please specify some for colonyID '6' Omitting from results."  Colony 7 has a
switchpoint at week 20 and therefore should restult in a convergence error
and the warning, "Warning: search for optimal switchpoint did not converge
for colonyID '6'. Omitting from results."
}
\keyword{datasets}
