% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot.bumbldf}
\alias{autoplot.bumbldf}
\title{Plot observed and fitted results from bumbl()}
\usage{
\method{autoplot}{bumbldf}(bumbldf, colony = NULL)
}
\arguments{
\item{bumbldf}{a dataframe produced by \code{\link[=bumbl]{bumbl()}}}

\item{colony}{a character vector of colony IDs to plot}
}
\value{
invisibly returns a ggplot object
}
\description{
Plots observed (points) and fitted (red line) values from the model implemented by \code{bumbl()}, faceted by colony.
}
\examples{
set.seed(687)
colonyID_subset <- sample(bombus$colony, 10)
colony_subset <- bombus[bombus$colony \%in\% colonyID_subset, ]
results <- bumbl(colony_subset, colonyID = colony, t = week,
                 formula = mass ~ week)
autoplot(results)
}
