% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{bipm}
\alias{bipm}
\title{Integral Projection Model for Bumblebee Colony Growth}
\usage{
bipm(
  larv_surv = 0.980419283573345,
  dev_time_mean = 23.5670474993181,
  wkr_size_min = 2.5,
  wkr_size_max = 5.81,
  wkr_size_mean = 3.64789499281592,
  wkr_size_sd = 0.427668095606862,
  poln_cost = 0.1275805,
  prop_foraging = 1,
  wkr_mass_f = mass_func,
  wkr_surv_f = surv_func,
  p_forage_f = forage_func,
  p_poln_ret_f = poln_ret_func,
  trips_f = trips_func,
  poln_mass_f = poln_mass_func
)
}
\arguments{
\item{larv_surv}{Larval survival rate.}

\item{dev_time_mean}{Mean development time from egg to adult, in days.}

\item{wkr_size_min}{Minimum observed worker intertegular span (ITS), in mm.}

\item{wkr_size_max}{Maximum observed worker ITS, in mm.}

\item{wkr_size_mean}{Mean observed worker ITS, in mm.}

\item{wkr_size_sd}{Observed standard deviation of worker ITS.}

\item{poln_cost}{Grams of pollen needed to make a gram of new worker.  The
default was estimated as the mean mass of pollen per cell / mean worker
mass.}

\item{prop_foraging}{Proportion of workers allowed to forage.  When less than
1, the smallest \code{1 - prop_foraging} workers do not contribute resources to
recruitment of new larvae.}

\item{wkr_mass_f}{A function relating worker mass to worker ITS.}

\item{wkr_surv_f}{Worker survival as a function of worker ITS.}

\item{p_forage_f}{Probability of a worker making a foraging trip as a
function of ITS.}

\item{p_poln_ret_f}{Probability of a worker returning from a foraging trip
with pollen, as a function of ITS.}

\item{trips_f}{The average number of foraging trips per day as a function of
worker ITS.}

\item{poln_mass_f}{The mass of pollen returned per trip as a function of ITS,
in grams}
}
\value{
A list containing the full integral projection model and the colony
growth rate, lambda.
}
\description{
Builds an size- and age-based integral projection model to estimate bumblebee
colony growth using vital rates estimated from previously published empirical
work, as described in Kerr, et al. (\emph{in prep}).
}
\details{
Defaults for all functions are estimated from data and published in
Kerr et al. (\emph{in prep}). Parameters ending in \verb{_f} expect the name of a
function or an anonymous function.
}
\examples{
# Original model:
bipm()
# Effect of reducing larval survival:
bipm(larv_surv = 0.97)
}
\references{
Kerr, N. Z., Williams, N. M., and Crone, E. E. (2020). Optimal
life history strategies in a eusocial insect: the contribution of worker
size polymorphism during colony establishment and growth phases in
bumblebee colonies. Journal of Animal Ecology (\emph{in prep}).

Kerr NZ, Crone EE, Williams NM (2019) Integrating vital rates explains
optimal worker size for resource return by bumblebee workers. Funct Ecol
33:467–478. doi: 10.1111/1365-2435.13251
}
\seealso{
\code{\link{mass_func}}, \code{\link{surv_func}},
\code{\link{forage_func}}, \code{\link{poln_ret_func}},
\code{\link{trips_func}}, \code{\link{poln_mass_func}}
}
