% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{bipm}
\alias{bipm}
\title{Integral Projection Model for Bumblebees}
\usage{
bipm(
  larv_surv = 0.9804193,
  dev_time_mean = 23.56705,
  wkr_size_min = 2.5,
  wkr_size_max = 5.81,
  wkr_size_mean = 3.600687,
  wkr_size_sd = 0.4276681,
  wkr_surv_f = function(wkr_size) plogis(4.5000243 - 0.4956151 * wkr_size),
  p_poln_ret_f = function(wkr_size) plogis(7.9804084 + (-4.9121457 * wkr_size) +
    (0.6533555 * wkr_size^2)),
  p_forage_f = function(wkr_size) plogis(-2.392183 + 1.364705 * wkr_size),
  trips_f = function(wkr_size) exp(-10.4418003 + (5.6902411 * wkr_size) + (-0.6896561 *
    wkr_size^2)),
  poln_mass_f = function(wkr_size) exp(-5.7240368 + 0.2914442 * wkr_size),
  poln_per_cell = 0.016
)
}
\arguments{
\item{larv_surv}{larval survival rate}

\item{dev_time_mean}{mean development time from egg to adult, in days}

\item{wkr_size_min}{minimum observed worker size}

\item{wkr_size_max}{maximum observed worker size}

\item{wkr_size_mean}{mean observed worker size}

\item{wkr_size_sd}{standard deviation of worker size}

\item{wkr_surv_f}{worker survival as a function of size}

\item{p_poln_ret_f}{probability of returning with pollen as a function of size}

\item{p_forage_f}{probability of making a foraging trip as a function of size}

\item{trips_f}{number of trips per day as a function of size}

\item{poln_mass_f}{the mass of pollen returned per trip as a function of size, in grams}

\item{poln_per_cell}{mean mass of pollen per cell, in grams}
}
\value{
the full integral projection matrix (invisibly)
}
\description{
Integral Projection Model for Bumblebees
}
\examples{
# Original model:
bipm()
# Effect of reducing larval survival:
bipm(larv_surv = 0.97)
}
