% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{brkpt}
\alias{brkpt}
\title{Fit breakpoint model to individual colony}
\usage{
brkpt(data, taus, t, formula)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{taus}{a vector of taus to test}

\item{t}{the unquoted variable representing time in ___units?}

\item{formula}{a formula passed to \code{lm}}
}
\value{
a tibble with a column for the winning tau and a column for the winning model
}
\description{
Fits models using a range of taus and picks the best one using maximum liklihood
}
\examples{
data(colony_weights)
testbees <- colony_weights[(colony_weights$ColonyID == 18), ]
mytaus <- (seq(2,8,0.1))
brkpt(testbees, mytaus, Round, log(TrueColonyWt_g) ~ Round)
}
