% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{brkpt}
\alias{brkpt}
\title{Fit breakpoint model to individual colony}
\usage{
brkpt(data, taus = NULL, t, formula, family = c("gaussian", "poisson",
  "overdispersed"))
}
\arguments{
\item{data}{a dataframe or tibble}

\item{taus}{an optional vector of taus to test. If not supplied, \code{seq(min(t), max(t), length.out = 50)} will be used}

\item{t}{the unquoted column name for the time variable in \code{data}}

\item{formula}{a formula passed to \code{lm}.  This should include the time variable supplied to \code{t}}

\item{family}{the model family to use.  By default ("gaussian") \code{lm()} is used.  "poisson" will run \code{glm(...family = poission(link = "log"))} and "overdispersed" will run \code{glm.nb} from the \code{MASS} package.  For continuous measures of colony growth such as mass, use \code{family = "gaussian"} (see example).  For count data such as number of workers, use the untransformed response variable and \code{family = "poisson"}.  For overdispersed count data, use \code{family = "overdispersed"}.}
}
\value{
a tibble with a column for the winning tau and a column for the winning model
}
\description{
Fits models using a range of taus and picks the best one using maximum liklihood
}
\examples{
testbees <- bombus[bombus$colony == 9, ]
# Using dates
brkpt(testbees, t = date, formula = log(mass) ~ date)
# Using weeks
brkpt(testbees, t = week, formula = log(mass) ~ week)
}
