% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{brkpt}
\alias{brkpt}
\alias{brkpt.nb}
\title{Fit breakpoint model to individual colony}
\usage{
brkpt(data, t, formula, family = gaussian(link = "log"), taus = NULL, ...)

brkpt.nb(data, taus = NULL, t, formula, link = "log", ...)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{t}{the unquoted column name for the time variable in \code{data}}

\item{formula}{a formula passed to \code{glm()} or \code{glm.nb()}.  This should
include the time variable supplied to \code{t}}

\item{family}{passed to \code{glm()}}

\item{taus}{an optional vector of taus to test. If not supplied, \code{seq(min(t), max(t), length.out = 50)} will be used}

\item{...}{additional arguments passed to \code{glm()} or \code{glm.nb()}}

\item{link}{passed to \code{glm.nb()}}
}
\value{
a tibble with a column for the winning tau and a column for the
winning model
}
\description{
Fits models using a range of taus and picks the best one using maximum
liklihood
}
\section{Functions}{
\itemize{
\item \code{brkpt.nb}: passes arguments to \code{MASS::glm.nb()}
}}

\examples{
testbees <- bombus[bombus$colony == 9, ]
# Using dates
\dontrun{
brkpt(testbees, t = date, formula = mass ~ date)
# Using weeks
brkpt(testbees, t = week, formula = mass ~ week)
}
}
\keyword{internal}
