% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{brkpt}
\alias{brkpt}
\title{Fit breakpoint model to individual colony}
\usage{
brkpt(data, taus = NULL, t, formula, family = c("gaussian", "poisson",
  "negbin"))
}
\arguments{
\item{data}{a dataframe or tibble}

\item{taus}{an optional vector of taus to test. If not supplied, \code{seq(min(t), max(t), length.out = 50)} will be used}

\item{t}{the unquoted column name for the time variable in \code{data}}

\item{formula}{a formula passed to \code{glm}.  This should include the time
variable supplied to \code{t}}

\item{family}{the model family to use.  By default, the data are fit with a
log-link gaussian generalized linear model. Because a log link is used, the
response variable should not be log-transformed.  For count data (e.g.
number of workers), use "poisson".  For overdispersed count data, use
"negbin" to fit the model using \code{glm.nb()} from the \code{MASS} package.}
}
\value{
a tibble with a column for the winning tau and a column for the
winning model
}
\description{
Fits models using a range of taus and picks the best one using maximum
liklihood
}
\examples{
testbees <- bombus[bombus$colony == 9, ]
# Using dates
brkpt(testbees, t = date, formula = mass ~ date)
# Using weeks
brkpt(testbees, t = week, formula = mass ~ week)
}
