% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony-growth.R
\name{bumbl}
\alias{bumbl}
\title{Fit breakpoint growth models to many colonies.
Fits breakpoint growth models to each colony in a dataset and returns the estimated breakpoint (tau) and some other things...}
\usage{
bumbl(data, colonyID, taus, t, formula)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{colonyID}{the unquoted column name of the colony ID variable}

\item{taus}{vector of taus to test}

\item{t}{the unquoted column name of the time variable in (units???)}

\item{formula}{a formula passed to `lm()`}
}
\value{
a dataframe
}
\description{
Fit breakpoint growth models to many colonies.
Fits breakpoint growth models to each colony in a dataset and returns the estimated breakpoint (tau) and some other things...
}
\examples{
data(colony_weights)
mytaus <- (seq(2,8,0.1))
bumbl(colony_weights, ColonyID, mytaus, Round, log(TrueColonyWt_g) ~ Round)
}
