% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colony_weights}
\alias{colony_weights}
\title{Bumblebee colony weight data from Crone and Williams, 2016}
\format{A data frame with 178 rows and 10 columns
\describe{
\item{Pair}{???}
\item{Site}{Site name}
\item{Round}{Which round of measurement (in 10 day increments)}
\item{ColonyID}{Colony identifier}
\item{TRT}{Early supplemented (ES) or late supplemented (LS)}
\item{Condition}{???}
\item{TrueColonyWt_g}{True colony weight, in grams}
\item{NectarWt_g}{The weight of nectar in the colonies, in grams}
\item{A_ColonyWt_g}{???}
\item{B_BoxWt_g}{The tare weight of the box the colony is in, in grams}
}}
\source{
\url{link to archived data???}
}
\usage{
colony_weights
}
\description{
A dataset containing the weights of bumblebee colonies...
}
\keyword{datasets}
