% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.bumbldf}
\alias{plot.bumbldf}
\title{Plot observed and fitted results from bumbl()}
\usage{
\method{plot}{bumbldf}(bumbldf)
}
\arguments{
\item{bumbldf}{a dataframe produced by \code{bumbl()}}
}
\value{
invisibly returns a list of dataframes used for building the plots.
}
\description{
Creates one plot per level of colonyID showing the observed (points) and fitted (red line) values from the model implemented by \code{bumbl()}.
}
\examples{
set.seed(687)
colonyID_subset <- sample(bombus$colony, 10)
colony_subset <- bombus[bombus$colony \%in\% colonyID_subset, ]
results <- bumbl(colony_subset, colonyID = colony, t = week,
                 formula = mass ~ week)
plot(results)
}
