% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{poln_mass_func}
\alias{poln_mass_func}
\title{Pollen mass as a function of bumblebee worker size}
\usage{
poln_mass_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of pollen mass in grams
}
\description{
The estimated pollen mass returned per foraging trip by a foraging worker
bumblebee as a function of body size (intertegular span) as described in Kerr
et al. 2019.
}
\examples{
poln_mass_func(4)
}
\references{
Kerr NZ, Crone EE, Williams NM (2019) Integrating vital rates
explains optimal worker size for resource return by bumblebee workers.
Funct Ecol 33:467–478. doi: 10.1111/1365-2435.13251
}
