% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{surv_func}
\alias{surv_func}
\title{Bumblebee worker survival as a function of size}
\usage{
surv_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of survival probabilities
}
\description{
Survival probabilities as a function of body size (intertegular span) for
bumblebee workers described in Kerr et al. 2019. Specifically, p =
plogis(4.50002431428571 - 0.495615057142857 * wkr_size)
}
\examples{
surv_func(4)
}
\references{
Kerr NZ, Crone EE, Williams NM (2019) Integrating vital rates
explains optimal worker size for resource return by bumblebee workers.
Funct Ecol 33:467–478. doi: 10.1111/1365-2435.13251
}
