% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{surv_func}
\alias{surv_func}
\title{Bumblebee worker survival as a function of size}
\usage{
surv_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of survival probabilities
}
\description{
Bumblebee worker survival as a function of size
}
\examples{
surv_func(4)
}
\references{
Kerr NZ, Crone EE, Williams NM (2019) Integrating vital rates
explains optimal worker size for resource return by bumblebee workers.
Funct Ecol 33:467–478. doi: 10.1111/1365-2435.13251
}
