% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{trips_func}
\alias{trips_func}
\title{Number of foraging trips per day as a function of bumblebee worker size}
\usage{
trips_func(wkr_size)
}
\arguments{
\item{wkr_size}{a vector of worker ITS values in mm}
}
\value{
a vector of mean number of trips per day
}
\description{
Number of foraging trips per day as a function of bumblebee worker size
}
\examples{
trips_func(4)
}
\references{
TODO: cite paper this comes from
}
